requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.9
reportlab==4.4.2
gunicorn